updates adoption from ed027b4474 to 75b42ff6d8

+++ core/src/test/java/ai/greycos/solver/core/impl/solver/termination/BestScoreFeasibleTerminationTest.java
@@ -6,10 +6,10 @@ import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import ai.greycos.solver.core.api.score.buildin.bendable.BendableScore;
-import ai.greycos.solver.core.api.score.buildin.hardsoft.HardSoftScore;
+import ai.greycos.solver.core.api.score.BendableScore;
+import ai.greycos.solver.core.api.score.HardSoftScore;
 import ai.greycos.solver.core.impl.phase.scope.AbstractPhaseScope;
-import ai.greycos.solver.core.impl.score.buildin.HardSoftScoreDefinition;
+import ai.greycos.solver.core.impl.score.definition.HardSoftScoreDefinition;
 import ai.greycos.solver.core.impl.score.definition.ScoreDefinition;
 import ai.greycos.solver.core.impl.score.director.InnerScore;
 import ai.greycos.solver.core.impl.solver.scope.SolverScope;
@@ -116,8 +116,8 @@ class BestScoreFeasibleTerminationTest {
     assertThat(
             termination.calculateFeasibilityTimeGradient(
                 InnerScore.fullyAssigned(
-                    BendableScore.of(new int[] {-10, -100}, new int[] {-50, -60, -70})),
-                BendableScore.of(new int[] {-4, -40}, new int[] {-50, -60, -70})))
+                    BendableScore.of(new long[] {-10, -100}, new long[] {-50, -60, -70})),
+                BendableScore.of(new long[] {-4, -40}, new long[] {-50, -60, -70})))
         .isEqualTo(0.6, offset(0.0));
   }
 }
